pre-commit + entrypt
